+++ lib/src/views/bookmarks/bookmarks_screen.dart
@@ -3,6 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:app_9news/src/provider/news_provider.dart';
 import 'package:app_9news/src/configs/app_routes.dart';
 import 'package:app_9news/src/views/homepage.dart'; // Impor LatestNewsCard
+import 'package:app_9news/src/widgets/news_cards.dart';
 
 class BookmarksScreen extends StatelessWidget {
   const BookmarksScreen({super.key});
