+++ lib/src/views/explore/explore_screen.dart
@@ -3,7 +3,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:app_9news/src/provider/news_provider.dart';
-import 'package:app_9news/src/views/homepage.dart'; // Impor untuk LatestNewsCard
+import 'package:app_9news/src/views/homepage.dart'; // Impor untuk Lates
+import 'package:app_9news/src/widgets/news_cards.dart';
 
 class ExploreScreen extends StatefulWidget {
   const ExploreScreen({super.key});
